requirements.txt
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Markdown==3.2.1